Move common irreducible/primitive poly functions to `_poly.py`
